+++ app/static/lib/editor.js
@@ -10,7 +10,11 @@ import { loadFacsimile } from './facsimile.js';
 import Viewer from './viewer.js';
 import { handleEditorChanges, version, versionDate } from './main.js';
 
-// smart indent selected region in editor, if none, do all
+/**
+ * Smart indents selected region in editor, if none, do all
+ * @param {Viewer} v
+ * @param {CodeMirror} cm
+ */
 export function indentSelection(v, cm) {
   v.updateNotation = false;
   cm.blockChanges = true;
@@ -37,7 +41,13 @@ export function indentSelection(v, cm) {
   v.updateNotation = true;
 } // indentSelection()
 
-// delete selected elements
+/**
+ * Deletes selected elements
+ * @param {Viewer} v
+ * @param {CodeMirror} cm
+ * @param {boolean} modifyerKey
+ * @returns
+ */
 export function deleteElement(v, cm, modifyerKey = false) {
   v.loadXml(cm.getValue(), true);
   let id = v.selectedElements[0]; // TODO: iterate over selectedElements
@@ -116,6 +126,15 @@ export function deleteElement(v, cm, modifyerKey = false) {
   v.updateNotation = true;
 } // deleteElement()
 
+/**
+ * Inserts a new control element to DOM and editor
+ * @param {Viewer} v
+ * @param {CodeMirror} cm
+ * @param {string} elName ('slur', 'dynam', ...)
+ * @param {string} placement ('below', 'up', ...)
+ * @param {string} form ('cres', 'inv', depending on element type)
+ * @returns
+ */
 export function addControlElement(v, cm, elName, placement, form) {
   if (v.selectedElements.length === undefined || v.selectedElements.length < 1) return;
   v.selectedElements = utils.sortElementsByScorePosition(v.selectedElements);
@@ -218,8 +237,7 @@ export function addControlElement(v, cm, elName, placement, form) {
     cm.replaceRange(dutils.xmlToString(newElement) + '\n', cm.getCursor());
     cm.indentLine(p1.line, 'smart');
     cm.indentLine(p1.line + 1, 'smart');
-    cm.setSelection(p1);
-    console.log('Cursor pos: ', cm.getCursor());
+    utils.setCursorToId(cm, uuid); // to select new element
   }
   // add new element to DOM
   var measureId = startEl.closest('measure').getAttribute('xml:id');
@@ -228,11 +246,19 @@ export function addControlElement(v, cm, elName, placement, form) {
   v.selectedElements = [];
   v.selectedElements.push(uuid);
   addApplicationInfo(v, cm);
-  addApplicationInfo(v, cm);
   v.updateData(cm, false, true);
   v.updateNotation = true;
 } // addControlElement()
 
+/**
+ * Inserts a clef change element before/after the current selection
+ * @param {Viewer} v
+ * @param {CodeMirror} cm
+ * @param {string} shape ('G', 'F', 'C')
+ * @param {string} line ('2', ...)
+ * @param {boolean} before element, or after (false)
+ * @returns
+ */
 export function addClefChange(v, cm, shape = 'G', line = '2', before = true) {
   if (v.selectedElements.length === 0) return;
   v.updateNotation = false; // stop update notation
@@ -255,17 +281,23 @@ export function addClefChange(v, cm, shape = 'G', line = '2', before = true) {
     cm.execCommand('goLineEnd');
     cm.replaceRange('\n' + dutils.xmlToString(newElement), cm.getCursor());
   }
-  cm.execCommand('indentAuto');
+  cm.execCommand('indentAuto'); // auto indent current line or selection
+  utils.setCursorToId(cm, uuid); // to select new element
   v.updateNotation = true; // update notation again
   v.selectedElements = [];
   v.selectedElements.push(uuid);
   v.lastNoteId = uuid;
   v.updatePage(cm, '', uuid);
-}
+} // addClefChange()
 
-// Reverse or insert att:placement (artic, ...), att.curvature (slur, tie,
-// phrase) and att.stems (note, chord) of current element
-// (or its children, such as all notes/chords within a beam).
+/**
+ * Reverses or inserts att:placement (artic, ...), att.curvature (slur, tie,
+ * phrase) and att.stems (note, chord) of current element
+ * (or its children, such as all notes/chords within a beam).
+ * @param {Viewer} v
+ * @param {CodeMirror} cm
+ * @param {boolean} modifier
+ */
 export function invertPlacement(v, cm, modifier = false) {
   v.loadXml(cm.getValue());
   let ids = utils.sortElementsByScorePosition(v.selectedElements);
@@ -301,23 +333,39 @@ export function invertPlacement(v, cm, modifier = false) {
         val = 'below';
       }
       if (modifier) {
-        let staffList = getStaffNumbersForClosestStaffGroup(v, el);
-        if (staffList.length === 2) {
-          if (el.hasAttribute(attr)) {
-            if (['above', 'below'].includes(el.getAttribute(attr))) {
-              val = 'between';
-              el.setAttribute('staff', staffList.sort().join(' '));
+        let response = getStaffNumbersForClosestStaffGroup(v, el);
+        if (response) {
+          let staffNumbers = []; // relevant two staves for @place='between'
+          if (response.staffNumbers.length === 2) {
+            staffNumbers = response.staffNumbers.sort();
+          } else {
+            // try to guess the relevant two staves
+            staffNumbers = [response.staffNumber];
+            const i = response.staffNumbers.indexOf(response.staffNumber);
+            if (i === response.staffNumbers.length - 1) {
+              staffNumbers.push(response.staffNumbers[i - 1]);
             } else {
-              val = 'above';
-              el.setAttribute('staff', staffList[0]);
+              staffNumbers.push(response.staffNumbers[i + 1]);
             }
+            let msg =
+              'Editor between placement: Please check staff numbers of ' +
+              el.nodeName +
+              ' (' +
+              id +
+              ') ' +
+              ' as it does not sit in a staff group with two staves' +
+              ' and relevant staves cannot be clearly determined.';
+            console.log(msg);
+            v.showAlert(msg, 'info');
           }
-        } else {
-          let msg =
-            'Editor between placement: Cannot change placement to "between", as selected element does not sit in a staff group with two staves.';
-          console.log(msg);
-          v.showAlert(msg, 'warning');
-          return;
+          // set @place and @staff attribute
+          if (!el.hasAttribute(attr) || (el.hasAttribute(attr) && el.getAttribute(attr) !== 'between')) {
+            val = 'between'; // set to between, if no or other @place attribute
+            el.setAttribute('staff', staffNumbers.sort().join(' '));
+          } else {
+            val = 'above'; // default value
+            el.setAttribute('staff', staffNumbers[0]);
+          }
         }
       }
       // for fermata, change form from inv to nothing or back
@@ -391,14 +439,18 @@ export function invertPlacement(v, cm, modifier = false) {
     }
   }
   // console.info('TextCursor: ', txtEdr.getCursorBufferPosition());
-  if (range) cm.setCursor(range.end);
   v.selectedElements = ids;
   addApplicationInfo(v, cm);
   v.updateData(cm, false, true);
   v.updateNotation = true; // update notation again
 } // invertPlacement()
 
-// toggle (switch on/off) artic to selected elements
+/**
+ * Toggles (switches on/off) artic at selected elements
+ * @param {Viewer} v
+ * @param {CodeMirror} cm
+ * @param {string} artic ('stacc', ...)
+ */
 export function toggleArtic(v, cm, artic = 'stacc') {
   v.loadXml(cm.getValue());
   let ids = speed.filterElements(v.selectedElements, v.xmlDoc);
@@ -417,7 +469,7 @@ export function toggleArtic(v, cm, artic = 'stacc') {
     let uuid;
     let noteList;
     if (['note', 'chord'].includes(note.nodeName)) {
-      uuid = toggleArticForNote(v, note, artic);
+      uuid = toggleArticForNote(note, artic, v.xmlIdStyle);
       uuid ? (ids[i] = uuid) : (ids[i] = id);
       range = replaceInEditor(cm, note, true);
       cm.execCommand('indentAuto');
@@ -425,21 +477,25 @@ export function toggleArtic(v, cm, artic = 'stacc') {
       let noteId;
       for (noteId of noteList) {
         note = v.xmlDoc.querySelector("[*|id='" + noteId + "']");
-        uuid = toggleArticForNote(v, note, artic);
+        uuid = toggleArticForNote(note, artic, v.xmlIdStyle);
         range = replaceInEditor(cm, note, true);
         cm.execCommand('indentAuto');
       }
     }
   }
-  if (range) cm.setCursor(range.end);
   v.selectedElements = ids;
   addApplicationInfo(v, cm);
   v.updateData(cm, false, true);
   v.updateNotation = true; // update notation again
 } // toggleArtic()
 
-// shift element (rests, note) up/down by pitch name (1 or 7 steps)
-export function shiftPitch(v, cm, deltaPitch) {
+/**
+ * Shifts element (rests, note) up/down by pitch name (1 or 7 steps)
+ * @param {Viewer} v
+ * @param {CodeMirror} cm
+ * @param {int} deltaPitch (-1, -12, +2)
+ */
+export function shiftPitch(v, cm, deltaPitch = 0) {
   v.loadXml(cm.getValue());
   let ids = speed.filterElements(v.selectedElements, v.xmlDoc);
   v.updateNotation = false;
@@ -451,9 +507,9 @@ export function shiftPitch(v, cm, deltaPitch) {
     let chs = Array.from(el.querySelectorAll('note,rest,mRest,multiRest'));
     if (chs.length > 0)
       // shift many elements
-      chs.forEach((ele) => replaceInEditor(cm, pitchMover(ele, deltaPitch)));
+      chs.forEach((ele) => replaceInEditor(cm, pitchMover(ele, deltaPitch)), true);
     // shift one element
-    else replaceInEditor(cm, pitchMover(el, deltaPitch));
+    else replaceInEditor(cm, pitchMover(el, deltaPitch), true);
   }
   v.selectedElements = ids;
   addApplicationInfo(v, cm);
@@ -461,6 +517,12 @@ export function shiftPitch(v, cm, deltaPitch) {
   v.updateNotation = true; // update notation again
 } // shiftPitch()
 
+/**
+ * Moves selected elements to next staff (without checking boundaries)
+ * @param {Viewer} v
+ * @param {CodeMirror} cm
+ * @param {boolean} upwards
+ */
 export function moveElementToNextStaff(v, cm, upwards = true) {
   console.info('moveElementToNextStaff(' + (upwards ? 'up' : 'down') + ')');
   v.loadXml(cm.getValue());
@@ -489,7 +551,13 @@ export function moveElementToNextStaff(v, cm, upwards = true) {
   v.updateNotation = true; // update notation again
 } // moveElementToNextStaff()
 
-// add beam, only speed mode
+/**
+ * Adds beam element
+ * @param {Viewer} v
+ * @param {CodeMirror} cm
+ * @param {string} elementName
+ * @returns
+ */
 export function addBeamElement(v, cm, elementName = 'beam') {
   v.loadXml(cm.getValue());
   v.selectedElements = speed.filterElements(v.selectedElements, v.xmlDoc);
@@ -542,7 +610,12 @@ export function addBeamElement(v, cm, elementName = 'beam') {
   v.updateNotation = true; // update notation again
 } // addBeamElement()
 
-// add beamSpan element
+/**
+ * Adds beamSpan element to selected elements
+ * @param {Viewer} v
+ * @param {CodeMirror} cm
+ * @returns
+ */
 export function addBeamSpan(v, cm) {
   v.loadXml(cm.getValue());
   if (v.selectedElements.length < 1) return;
@@ -574,7 +647,7 @@ export function addBeamSpan(v, cm) {
     cm.replaceRange(dutils.xmlToString(beamSpan) + '\n', cm.getCursor());
     cm.indentLine(p1.line, 'smart'); // TODO
     cm.indentLine(p1.line + 1, 'smart');
-    cm.setSelection(p1);
+    utils.setCursorToId(cm, uuid);
   }
   v.selectedElements = [];
   v.selectedElements.push(uuid);
@@ -584,7 +657,14 @@ export function addBeamSpan(v, cm) {
   v.updateNotation = true; // update notation again
 } // addBeamSpan()
 
-// add octave element and modify notes inside selected elements
+/**
+ * Adds an octave element and modifies notes inside selected elements
+ * @param {Viewer} v
+ * @param {CodeMirror} cm
+ * @param {string} disPlace ('above', 'below')
+ * @param {string} dis (8, 15, 22)
+ * @returns
+ */
 export function addOctaveElement(v, cm, disPlace = 'above', dis = '8') {
   v.loadXml(cm.getValue());
   if (v.selectedElements.length < 1) return;
@@ -610,7 +690,7 @@ export function addOctaveElement(v, cm, disPlace = 'above', dis = '8') {
     cm.replaceRange(dutils.xmlToString(octave) + '\n', cm.getCursor());
     cm.indentLine(p1.line, 'smart'); // TODO
     cm.indentLine(p1.line + 1, 'smart');
-    cm.setSelection(p1);
+    utils.setCursorToId(cm, uuid);
   }
   // find plist and modify elements
   findAndModifyOctaveElements(cm, v.xmlDoc, id1, id2, disPlace, dis);
@@ -623,8 +703,20 @@ export function addOctaveElement(v, cm, disPlace = 'above', dis = '8') {
   v.updateNotation = true; // update notation again
 } // addOctaveElement()
 
-// surround selected elements with a supplied element (and a responsibility
-// statement from v.respId
+/**
+ * Surrounds selected elements with a supplied element
+ * (and a responsibility statement from v.respId derived
+ * from mei-friend settings.
+ *
+ * If attrName is specified, it searches for those attributes,
+ * inserts them as new elements, and surrounds them with supplied
+ * elements. If there is already such a artic/accid child element,
+ * take it and surround it.
+ * @param {Viewer} v
+ * @param {CodeMirror} cm
+ * @param {string} attrName ('artic', 'accid')
+ * @returns
+ */
 export function addSuppliedElement(v, cm, attrName = 'none') {
   v.loadXml(cm.getValue());
   v.selectedElements = speed.filterElements(v.selectedElements, v.xmlDoc);
@@ -643,22 +735,33 @@ export function addSuppliedElement(v, cm, attrName = 'none') {
       // convert attrName to element and surround that
       if (attrName === 'artic' || attrName === 'accid') {
         if (!el.hasAttribute(attrName)) {
-          const msg = 'No ' + attrName + ' attribute in element ' + el.nodeName + '.';
-          console.log(msg);
-          v.showAlert(msg, 'warning');
-          return;
+          let childElement;
+          el.childNodes.forEach((ch) => {
+            if (ch.nodeName === attrName) childElement = ch;
+          });
+          if (childElement) {
+            parent = el;
+            el = childElement;
+          } else {
+            const msg =
+              'No ' + attrName + ' attribute or child node found in element ' + el.nodeName + ' (' + id + ').';
+            console.log(msg);
+            v.showAlert(msg, 'warning');
+            return;
+          }
+        } else {
+          let attrValue = el.getAttribute(attrName);
+          let attrEl = document.createElementNS(dutils.meiNameSpace, attrName);
+          let uuid = mintSuppliedId(id, attrName);
+          attrEl.setAttributeNS(dutils.xmlNameSpace, 'xml:id', uuid);
+          attrEl.setAttribute(attrName, attrValue);
+          el.removeAttribute(attrName);
+          el.appendChild(attrEl);
+          replaceInEditor(cm, el, true);
+          cm.execCommand('indentAuto');
+          parent = el;
+          el = attrEl;
         }
-        let attrValue = el.getAttribute(attrName);
-        let attrEl = document.createElementNS(dutils.meiNameSpace, attrName);
-        let uuid = mintSuppliedId(id, attrName);
-        attrEl.setAttributeNS(dutils.xmlNameSpace, 'xml:id', uuid);
-        attrEl.setAttribute(attrName, attrValue);
-        el.removeAttribute(attrName);
-        el.appendChild(attrEl);
-        replaceInEditor(cm, el, true);
-        cm.execCommand('indentAuto');
-        parent = el;
-        el = attrEl;
       }
 
       let sup = document.createElementNS(dutils.meiNameSpace, 'supplied');
@@ -689,6 +792,12 @@ export function addSuppliedElement(v, cm, attrName = 'none') {
   }
 } // addSuppliedElement()
 
+/**
+ * Adds a vertical group attribute (vgrp) to all selected elements
+ * @param {Viewer} v
+ * @param {CodeMirror} cm
+ * @returns
+ */
 export function addVerticalGroup(v, cm) {
   v.loadXml(cm.getValue());
   v.selectedElements = speed.filterElements(v.selectedElements, v.xmlDoc);
@@ -724,7 +833,7 @@ export function addVerticalGroup(v, cm) {
 /**
  * Adds an application element to appInfo or updates its date, if already there
  * @param {Viewer} v
- * @param {object} cm
+ * @param {CodeMirror} cm
  */
 export function addApplicationInfo(v, cm) {
   if (document.getElementById('addApplicationNote').checked) {
@@ -785,7 +894,12 @@ export function addApplicationInfo(v, cm) {
   }
 } // addApplicationInfo()
 
-// wrapper for cleaning superfluous @accid.ges attributes
+/**
+ * Wrapper function to utils.cleanAccid() for cleaning
+ * superfluous @accid.ges attributes
+ * @param {Viewer} v
+ * @param {CodeMirror} cm
+ */
 export function cleanAccid(v, cm) {
   v.updateNotation = false;
   v.loadXml(cm.getValue(), true);
@@ -793,8 +907,13 @@ export function cleanAccid(v, cm) {
   v.updateNotation = true;
 }
 
-// wrapper for renumbering measure numberlike string
-export function renumberMeasures(v, cm, change) {
+/**
+ * Wrapper function for renumbering measure numberlike attribute (@n)
+ * @param {Viewer} v
+ * @param {CodeMirror} cm
+ * @param {boolean} change
+ */
+export function renumberMeasures(v, cm, change = false) {
   v.updateNotation = false;
   v.loadXml(cm.getValue(), true);
   utils.renumberMeasures(v, cm, 1, change);
@@ -804,7 +923,12 @@ export function renumberMeasures(v, cm, change) {
   v.updateNotation = true;
 } // renumberMeasures()
 
-// function for adding/removing xml:ids in xmlDoc & reloading MEI into cm
+/**
+ * Function for adding/removing xml:ids in xmlDoc & reloading MEI into cm
+ * @param {Viewer} v
+ * @param {CodeMirror} cm
+ * @param {boolean} removeIds
+ */
 export function manipulateXmlIds(v, cm, removeIds = false) {
   let startTime = Date.now();
   let report = { added: 0, removed: 0 };
@@ -877,7 +1001,7 @@ export function manipulateXmlIds(v, cm, removeIds = false) {
 } // manipulateXmlIds()
 
 /**
- * Add zone element in editor (called from source-imager.js),
+ * Adds a zone element in editor (called from source-imager.js),
  * places it
  * @param {Viewer} v
  * @param {CodeMirror} cm
@@ -1000,7 +1124,14 @@ export function addZone(v, cm, rect, addMeasure = true) {
   }
 } // addZone()
 
-// remove zone in editor, called from editor.js
+/**
+ * Removes a zone in editor, called from editor.js
+ * @param {Viewer} v
+ * @param {CodeMirror} cm
+ * @param {Element} zone
+ * @param {boolean} removeMeasure
+ * @returns
+ */
 export function removeZone(v, cm, zone, removeMeasure = false) {
   if (!zone) return;
   removeInEditor(cm, zone);
@@ -1095,7 +1226,11 @@ export function addFacsimile(v, cm) {
   v.updateNotation = true;
 } // addFacsimile()
 
-// find xmlNode in textBuffer and remove it (including empty line)
+/**
+ * Finds xmlNode in textBuffer and removes it (including empty line)
+ * @param {CodeMirror} cm
+ * @param {Element} xmlNode
+ */
 export function removeInEditor(cm, xmlNode) {
   let itemId = xmlNode.getAttribute('xml:id');
   let searchSelfClosing = '(?:<' + xmlNode.nodeName + `)(\\s+?)([^>]*?)(?:xml:id=["']` + itemId + `['"])([^>]*?)(?:/>)`;
@@ -1128,13 +1263,20 @@ export function removeInEditor(cm, xmlNode) {
       }
     }
   } else console.info('removeInEditor(): nothing removed for ' + itemId + '.');
-}
+} // removeInEditor()
 
 function isEmpty(str) {
   return !/\S/g.test(str);
-}
+} // isEmpty()
 
-// find xmlNode in textBuffer and replace it with new serialized content
+/**
+ * Finds xmlNode in textBuffer and replaces it with new serialized content
+ * @param {CodeMirror} cm
+ * @param {Element} xmlNode
+ * @param {boolean} select
+ * @param {Element} newNode
+ * @returns
+ */
 export function replaceInEditor(cm, xmlNode, select = false, newNode = null) {
   let newMEI = newNode ? dutils.xmlToString(newNode) : dutils.xmlToString(xmlNode);
   // search in buffer
@@ -1176,13 +1318,20 @@ export function replaceInEditor(cm, xmlNode, select = false, newNode = null) {
     start: sc.from(),
     end: sc.to(),
   };
-}
+} // replaceInEditor()
 
 // ############################################################################
 // # (mostly) private functions                                               #
 // ############################################################################
 
-function toggleArticForNote(v, note, artic) {
+/**
+ * Toggles and articulation for a given element (note)
+ * @param {Element} note
+ * @param {string} artic
+ * @param {string} xmlIdStyle (Original, Base36, mei-friend, see viewer.js)
+ * @returns
+ */
+function toggleArticForNote(note, artic, xmlIdStyle) {
   note = utils.attrAsElements(note);
   let articChildren;
   let add = false;
@@ -1205,15 +1354,20 @@ function toggleArticForNote(v, note, artic) {
   if (add) {
     // add artic as element
     let articElement = document.createElementNS(dutils.meiNameSpace, 'artic');
-    uuid = utils.generateXmlId('artic', v.xmlIdStyle);
+    uuid = utils.generateXmlId('artic', xmlIdStyle);
     articElement.setAttributeNS(dutils.xmlNameSpace, 'xml:id', uuid);
     articElement.setAttribute('artic', artic);
     note.appendChild(articElement);
   }
-  // console.info('modified element: ', note);
   return uuid;
-}
+} // toggleArticForNote()
 
+/**
+ * Modifies an element's pitch up and down (i.e. manipulating @oct, @pname)
+ * @param {Element} el
+ * @param {number} deltaPitch
+ * @returns
+ */
 function pitchMover(el, deltaPitch) {
   let oct = 4;
   let pname = 'c';
@@ -1239,9 +1393,15 @@ function pitchMover(el, deltaPitch) {
   el.setAttribute(o, oct);
   el.setAttribute(p, att.pnames[pi]);
   return el;
-}
+} // pitchMover()
 
-function staffMover(cm, el, upwards) {
+/**
+ * Moves an element across staves.
+ * @param {CodeMirror} cm
+ * @param {Element} el
+ * @param {boolean} upwards
+ */
+function staffMover(cm, el, upwards = true) {
   let staff = el.closest('staff');
   let staffNo = -1;
   if (staff) staffNo = parseInt(staff.getAttribute('n'));
@@ -1261,12 +1421,22 @@ function staffMover(cm, el, upwards) {
   }
   if (staffNo === newStaffNo) el.removeAttribute('staff');
   else el.setAttribute('staff', newStaffNo);
-  replaceInEditor(cm, el);
-}
+  replaceInEditor(cm, el, true);
+} // staffMover()
 
-// find all notes between two ids in the same staff and set @oct.ges and
-// modify @oct with deltaOct (1,-1, 2,...);
-// or, if set=false: remove @oct.ges and reset @oct
+/**
+ * Finds all notes between two ids in the same staff and
+ * sets @oct.ges and modifies @oct with deltaOct (1,-1, 2,...);
+ * or, if set=false: remove @oct.ges and reset @oct
+ * @param {CodeMirror} cm
+ * @param {Document} xmlDoc
+ * @param {string} id1
+ * @param {string} id2
+ * @param {string} disPlace ('below|above')
+ * @param {string} dis (e.g. 8, 15, 22)
+ * @param {boolean} add
+ * @returns
+ */
 function findAndModifyOctaveElements(cm, xmlDoc, id1, id2, disPlace, dis, add = true) {
   let deltaOct = (parseInt(dis) - 1) / 7;
   if (disPlace === 'below') deltaOct *= -1; // normal logic: minus 1 when below
@@ -1315,11 +1485,11 @@ function findAndModifyOctaveElements(cm, xmlDoc, id1, id2, disPlace, dis, add = 
 } // findAndModifyOctaveElements()
 
 /**
- * Determines an array of staff numbers for a given element that spans
+ * Determines staff numbers for a given element that spans
  * the relevant staff group that the element is inside.
  * @param {Viewer} v
  * @param {Element} element
- * @returns {Array[]} staffNumbers
+ * @returns {Object} with staffNumbers and staffNumber as keys
  */
 function getStaffNumbersForClosestStaffGroup(v, element) {
   if (!element) return null;
@@ -1348,9 +1518,9 @@ function getStaffNumbersForClosestStaffGroup(v, element) {
           const n = st.getAttribute('n');
           if (n) staffNumbers.push(n);
         });
-        return staffNumbers;
+        return { staffNumber: staffNumber, staffNumbers: staffNumbers };
       }
     }
-    return [];
+    return { staffNumber: staffNumber, staffNumbers: [] };
   }
 } // findClosestStaffGroup()