+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Support for inserting supplied elements around artic/accid attributes (that get converted to elements) (fix for #14)
 * Do not save and insert slur below with `CTRL + S` on Windows/Linux (fix for #15)
 * Support for `@place="between"` on control elements (fix for #16)
+* Several improvements and harmonization in editor code
 
 ### 0.8.2 patch
 * Fixes to highlight synchronization edge-cases caused by rounding issues